% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_in_string.R
\name{find_in_string}
\alias{find_in_string}
\title{Find a vector of strings/numbers in a single string.}
\usage{
find_in_string(b, a)
}
\arguments{
\item{b}{numeric or character vector, which includes all the elements you are looking for.}

\item{a}{the string to be searched}
}
\value{
boolean. True or False
}
\description{
\code{find_in_string} returns the sum of all the values present in its arguments.
}
\details{
searches for all elements of vector b in string a, and returns true/false depending on whether or not all the elements
in b can also be found in a, regardless of order.
}
\examples{
b<-c("i", "x", "v")
a<-"ix"
c <-"vix"
find_in_string(b,a)
find_in_string(b,c)
}
\keyword{keys,}
\keyword{multiple,}
\keyword{search}
\keyword{string,}
