% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_in_vector.R
\name{list_in_vector}
\alias{list_in_vector}
\title{Find a vector of strings/numbers in a single string.}
\usage{
list_in_vector(list, vector, all = FALSE)
}
\arguments{
\item{list}{list of either numeric or character vectors that you would like to find.}

\item{vector}{a numeric or character vector to be searched}

\item{all}{default is FALSE. If FALSE, returns a logical vector. If TRUE, returns a boolean indicating whether or not all elements contained matches.}
}
\value{
boolean. True or False
}
\description{
\code{list_in_vector} returns the sum of all the values present in its arguments.
}
\details{
Test whether each element of list[[index]] can be found in the corresponding vector[index] for the whole list. If yes, returns TRUE else returns FALSE
}
\examples{
list <- list(c('hi','bi','sc'),c('h','o','g'),c('fu','lc','rum'))
vector <-c('hibiscus','goh','fulrum')
list_in_vector(list,vector)
list_in_vector(list,vector,all=FALSE)
}
\keyword{keys,}
\keyword{list}
\keyword{multiple,}
\keyword{vector,}
