% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_in_vector.R
\name{vector_in_vector}
\alias{vector_in_vector}
\title{A Function for Finding Partial Matches in Vectors}
\usage{
vector_in_vector(needle, haystack, vector = TRUE)
}
\arguments{
\item{needle}{the vector whose elements are searched for}

\item{haystack}{the vector whose elements are searched}

\item{vector}{Logical. Default set to TRUE. Returns a logical vector for matches in each element of haystack. If vector = FALSE, a single boolean is returned indicating if all elements have a match.}
}
\description{
This function searches for each element of vector1 (needle) in the corresponding element of vector2 (haystack), and returns a logical vector. Unlike the %in% function, the search key can be found anywhere in element being searched (not just the beginning).
}
\examples{
needle <- c("i","ii","iii")
haystack <- c("i","xniiv", "ximini")
vector_in_vector(needle,haystack, vector = TRUE)
}
\keyword{find}
\keyword{match,}
\keyword{vector,}
